Isolate GPU function in engine.py for ZeroGPU detection
- Create engine.py with @spaces.GPU decorated function
- Import process_message from engine.py into app.py
- Remove uvicorn.run (Spaces provides its own runner)
- Add spaces>=0.28.3 to requirements.txt
- Fixes: No @spaces.GPU function detected during startup
