+++ engine.py
@@ -1,0 +1,48 @@
+"""GPU-accelerated processing engine for Arabic teaching.
+
+This module contains @spaces.GPU decorated functions.
+Isolated in separate file for reliable ZeroGPU detection.
+"""
+
+import uuid
+
+import spaces
+import torch
+
+
+@spaces.GPU(duration=60)
+def process_message(message, chat_history, session_id):
+    """Process user message with agent (GPU-accelerated).
+
+    Currently: Simple logic placeholder with GPU verification
+    Future: Will load Qwen 7B models for teaching/grading
+
+    Args:
+        message: User input text
+        chat_history: List of message dicts with 'role' and 'content'
+        session_id: Session identifier
+
+    Returns:
+        Tuple of (empty_string, updated_chat_history)
+    """
+    # Verify GPU access (will be used for model inference)
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+
+    # Generate session ID if not provided
+    if not session_id:
+        session_id = str(uuid.uuid4())[:8]
+
+    # Simple agent logic (placeholder for Qwen 7B inference)
+    if "hello" in message.lower():
+        response = f"مرحباً! Hello! (Running on {device})"
+    elif "vocab" in message.lower():
+        response = "📚 Vocabulary mode activated!"
+    elif "grammar" in message.lower():
+        response = "📖 Grammar mode activated!"
+    else:
+        response = f"Echo: {message}"
+
+    # Use messages format for Gradio
+    chat_history.append({"role": "user", "content": message})
+    chat_history.append({"role": "assistant", "content": response})
+    return "", chat_history
